Changed get_random_word to work with a proxy
This change makes the POST request take into account ENV['http_proxy'] variable.
This is important when the code runs behind a proxy. The post_form ignores
the proxy variable, the new method uses the proxy.

Also a comment is added for people wanting to test it right, as curl method
might not be enough.