Updating irb comment in hangperson_game.rb
Currently, irb is NOT in the gemfile. It either needs to be added to the gemfile and then the old command tested ($ bundle exec irb -I. -r app.rb) or you can simply direct those who are curious enough to try this step to install irb and run irb without bundle (as suggested here).